+++ examples/react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @0xsequence/kit-example-react
 
+## 0.7.14
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @0xsequence/kit@2.10.1
+  - @0xsequence/kit-checkout@2.10.1
+  - @0xsequence/kit-wallet@2.10.1
+
 ## 0.7.13
 
 ### Patch Changes